TASK: Extract compiled classes slot from CacheManager
