+++ Neos.Flow/Classes/Cache/CacheManager.php
@@ -19,7 +19,6 @@ use Neos\Cache\Frontend\FrontendInterface;
 use Neos\Cache\Frontend\VariableFrontend;
 use Neos\Flow\Configuration\ConfigurationManager;
 use Neos\Flow\Log\Utility\LogEnvironment;
-use Neos\Flow\Reflection\ReflectionService;
 use Neos\Flow\Utility\Environment;
 use Neos\Utility\Files;
 use Neos\Flow\Utility\PhpAnalyzer;
@@ -62,12 +61,6 @@ class CacheManager
      * @var array
      */
     protected $persistentCaches = [];
-
-    /**
-     * The ReflectionService is only needed during compile time for handling
-     * @var ReflectionService
-     */
-    protected $reflectionService;
 
     /**
      * @var array
@@ -411,42 +404,6 @@ class CacheManager
             $objectConfigurationCache->remove('allAspectClassesUpToDate');
             $objectConfigurationCache->remove('allCompiledCodeUpToDate');
             $objectClassesCache->flush();
-        }
-    }
-
-    /**
-     * A slot that flushes caches as needed if classes with specific annotations have changed
-     *
-     * @param array<string> $classNames The full class names of the classes that got compiled
-     * @return void
-     */
-    public function flushConfigurationCachesByCompiledClass(array $classNames): void
-    {
-        $caches = [
-            Neos\Flow\Annotations\Route::class => ['Flow_Mvc_Routing_Route', 'Flow_Mvc_Routing_Resolve']
-        ];
-        $cachesToFlush = [];
-
-        foreach ($classNames as $className) {
-            foreach ($caches as $annotationClass => $cacheNames) {
-                if (!$this->reflectionService->isClassAnnotatedWith($className, $annotationClass)
-                    && count($this->reflectionService->getMethodsAnnotatedWith($className, $annotationClass)) === 0) {
-                    continue;
-                }
-                foreach ($caches[$annotationClass] as $cacheName) {
-                    $cachesToFlush[$cacheName] = $annotationClass;
-                }
-            }
-        }
-
-        foreach ($cachesToFlush as $cacheName => $annotationClass) {
-            $this->logger->info(sprintf('A class file containing the annotation "%s" has been changed, flushing related cache "%s"', $annotationClass, $cacheName), LogEnvironment::fromMethodName(__METHOD__));
-            $this->getCache($cacheName)->flush();
-        }
-
-        if (count($cachesToFlush) > 0) {
-            $this->logger->info('An annotated class file has been changed, refreshing compiled configuration cache', LogEnvironment::fromMethodName(__METHOD__));
-            $this->configurationManager->refreshConfiguration();
         }
     }
