+++ Neos.Flow/Classes/Package.php
@@ -11,7 +11,7 @@ namespace Neos\Flow;
  * source code.
  */
 
-use Neos\Flow\Cache\CacheManager;
+use Neos\Flow\Cache\AnnotationsCacheFlusher;
 use Neos\Flow\Core\Booting\Step;
 use Neos\Flow\Http\Helper\SecurityHelper;
 use Neos\Flow\ObjectManagement\CompileTimeObjectManager;
@@ -149,6 +149,6 @@ class Package extends BasePackage
         $dispatcher->connect(AuthenticationProviderManager::class, 'successfullyAuthenticated', Context::class, 'refreshRoles');
         $dispatcher->connect(AuthenticationProviderManager::class, 'loggedOut', Context::class, 'refreshTokens');
 
-        $dispatcher->connect(Proxy\Compiler::class, 'compiledClasses', CacheManager::class, 'flushCachesByCompiledClass');
+        $dispatcher->connect(Proxy\Compiler::class, 'compiledClasses', AnnotationsCacheFlusher::class, 'flushCachesByCompiledClass');
     }
 }
